Moving RoleStorageProviderModel to the legacy modules
Closes #12656
