+++ model/legacy/src/main/java/org/keycloak/models/LegacyRealmModel.java
@@ -20,6 +20,8 @@ package org.keycloak.models;
 import org.keycloak.models.RealmModel;
 import org.keycloak.storage.client.ClientStorageProvider;
 import org.keycloak.storage.client.ClientStorageProviderModel;
+import org.keycloak.storage.role.RoleStorageProvider;
+import org.keycloak.storage.role.RoleStorageProviderModel;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -47,4 +49,24 @@ public interface LegacyRealmModel extends RealmModel {
                 .map(ClientStorageProviderModel::new)
                 .sorted(ClientStorageProviderModel.comparator);
     }
+
+    /**
+     * @deprecated Use {@link #getRoleStorageProvidersStream() getRoleStorageProvidersStream} instead.
+     */
+    @Deprecated
+    default List<RoleStorageProviderModel> getRoleStorageProviders() {
+        return getRoleStorageProvidersStream().collect(Collectors.toList());
+    }
+
+    /**
+     * Returns sorted {@link RoleStorageProviderModel RoleStorageProviderModel} as a stream.
+     * It should be used with forEachOrdered if the ordering is required.
+     * @return Sorted stream of {@link RoleStorageProviderModel}. Never returns {@code null}.
+     */
+    default Stream<RoleStorageProviderModel> getRoleStorageProvidersStream() {
+        return getComponentsStream(getId(), RoleStorageProvider.class.getName())
+                .map(RoleStorageProviderModel::new)
+                .sorted(RoleStorageProviderModel.comparator);
+    }
+
 }
